+++ Source/Core/General/CRC.cs
@@ -142,7 +142,7 @@ namespace CodeImp.DoomBuilder
 		};
 
 			internal static uint ComputeCrc32(uint oldCrc, byte value) {
-				return (uint)(Crc32.CrcTable[(oldCrc ^ value) & 0xFF] ^ (oldCrc >> 8));
+				return (Crc32.CrcTable[(oldCrc ^ value) & 0xFF] ^ (oldCrc >> 8));
 			}
 
 			/// <summary>
@@ -155,7 +155,7 @@ namespace CodeImp.DoomBuilder
 			/// </summary>
 			public long Value {
 				get {
-					return (long)crc;
+					return crc;
 				}
 				set {
 					crc = (uint)value;
