+++ Source/Core/IO/HexenMapSetIO.cs
@@ -203,7 +203,7 @@ namespace CodeImp.DoomBuilder.IO
 				y = reader.ReadInt16();
 
 				// Create new item
-				v = map.CreateVertex(new Vector2D((float)x, (float)y));
+				v = map.CreateVertex(new Vector2D(x, y));
 				
 				// Add it to the lookup table
 				link.Add(i, v);
