+++ Source/Plugins/BuilderModes/Interface/CurveLinedefsForm.cs
@@ -42,8 +42,8 @@ namespace CodeImp.DoomBuilder.BuilderModes
 		#region ================== Properties
 
 		public int Vertices { get { return -verticesbar.Value; } }
-		public float Distance { get { return (float)-distancebar.Value; } }
-		public float Angle { get { return Angle2D.DegToRad((float)-anglebar.Value); } }
+		public float Distance { get { return -distancebar.Value; } }
+		public float Angle { get { return Angle2D.DegToRad(-anglebar.Value); } }
 		public bool FixedCurve { get { return circular.Checked; } }
 		public bool Backwards { get { return backwards.Checked; } }
